get tatib on user
